Add files via upload
Hi AP, Stank recommended me to change the pets that had biological type so they become mechanical, all others properties are the same as initial except that change. This is in order to restore equality between pets: some could be hugged, some not. Very unfair and frustrating for players who think they only buy a pet skin (this was my case btw) especially due to the cost of pet "skin" (5K).