Reworked Charlie Company PATCH NOTES: Charlie company units escort random marines within nearest marine group. Except when too far, units will continue to escort the same marines.
ADDITIONAL NOTES:
Siege mode is implemented, which breaks the escort; thus,
a while-loop is used to check the positions of charlie company units,
and reassigns or resumes the escorts when applicable.
33% to 67% of tanks will morph into siege mode while within range
of 1.35 to 3.0 of any marine. At the same time,
33% to 66% of the remaining tanks will remain in tank mode.
As a result of these conditions, "hugging" a siege tank will cause
the tank to unsiege. This is useful to prevent body blocking.
Charlie company units escort marines within a circle
between radius 1.5 to 5.0.