Fixing dr tanaka dying after completing his lab trip.
Triggers responsible for penalizing team for not protecting dr.tanaka and trigger responsible for ending game as failure in case tanaka dies were being disabled a few seconds too early.
They got disabled when mission was completed but before dr. was made invincible.

This fix simply moves the actions a few lines down in trigger. To after doctor is invincible.