Movespeed bonus from execute kills now in data.
When execute successfully kills a target recon is granted a movespeed bonus, it was achieved in trigger. While small and simple the trigger was spliting functionality of this ability from pure data to data/trigger hybrid.
From gameplay perspective it behaves exactly the same. For instantly killed target it's now a simple apply behavior onto recon as those are guaranteed dead. For non insta kill target it adds a short damage response buff to target which if damage is fatal aplies the buff to recon.