Fixing issues with x1-bot not holding fire properly.
That is a somewhat strange bug. Holdfire on x1 robot was working only until robot started to move. The moment it moved, holdfire was ignored.
Reason for that is having a turret which is allowed to attack on the move - this turret aquires targets on it's own which caused it to ignore holdfire on bot.

This fix adds a hidden behavior which comes into play the moment bot gets issued 'holdfire' command.
This behavior modifies the flag which stops the bot/turret from aquiring targets unless commanded with 'attack-move'