Spammable infestor kill objective.
In case one of infestors failed to spawn the event in 'InfestorKill' pointed to null which made it work for any unit dying on map.
It's caused by me removing condition testing for "DeadUnit=Infestor" when optimizing events, but instead of re-adding the condition it's better idea to ensure infestors spawn properly - this pull does it.