In rare cases items dropped onto ground were not turning neutral but remained owned by the player. Examples of those situations are: a)dropping ammo when inventory of marine is full b)buying an item when inventory of marine is full
It had pretty insignificant impact on gameplay in form of:
a)items owned by player could be seen by him thru the fog of war
b)player who owned the item could not use 'LaserDesignate' on that item in case he was a recon

I'm changing it coz a player whose name I can't recall atm asked me for it multiple times with complaint that => "ammo i drop as recon cannot be LD'ed and it makes running civs in ec harder coz i was dropping the mag and LD'ing it to save up on running".
Also it makes it consistent with manually dropped items which turn neutral when dropped.